Parse operatingHours json in API class
